penumbra_sdk_keys
Module keys

Structs

AddressIndex

AuthorizationKeyVar

Bip44Path

Represents a BIP44 derivation path.

Diversifier

DiversifierKey

FullViewingKey

The root viewing capability for all data related to a given spend authority.

IncomingViewingKey

Allows viewing incoming notes, i.e., notes sent to the spending key this key is derived from.

IncomingViewingKeyVar

NullifierKey

Allows deriving the nullifier associated with a positioned piece of state.

NullifierKeyVar

Represents the NullifierKey as a variable in an R1CS constraint system.

OutgoingViewingKey

Allows viewing outgoing notes, i.e., notes sent from the spending key this key is derived from.

RandomizedVerificationKey

SeedPhrase

A mnemonic seed phrase. Used to generate [SpendSeed]s.

SpendAuthRandomizerVar

SpendKey

A key representing a single spending authority.

SpendKeyBytes

A refinement type for a [u8; 32] indicating that it stores the bytes of a spend key.

WalletId

The hash of a full viewing key, used as an account identifier.

Constants

DIVERSIFIER_LEN_BYTES

IVK_LEN_BYTES

NK_LEN_BYTES

OVK_LEN_BYTES

SPENDKEY_LEN_BYTES